requirements.txt
streamlit==0.86.0
pandas==1.3.1
numpy==1.19.5
urllib3==1.25.8
json5==0.9.1
jsonschema==3.2.0
plotly==5.1.0
Pillow==7.0.0
mpmath==1.1.0
pathlib2==2.3.5
requests==2.25.1